renamed commands.js to commandhandler.js
