+++ app.js
@@ -1,6 +1,6 @@
 const Discord = require("discord.js");
 const config = require("./config.json");
-const commandHandler = require("./commands.js");
+const commandHandler = require("./commandhandler.js");
 const imjoke = require("./imjoke.js");
 const chickenButt = require("./chickenbutt.js");
 const atSomeone = require("./atsomeone.js");
