+++ commandhandler.js
